refactor (dictionary-evaluation): split matcher evaluation into strategy-based modules
- replace monolithic candidate evaluator with five focused evaluation files:
  StandardEvaluation, SplitJoinEvaluation, MergedTokenEvaluation,
  ThreeTokenEvaluation, and EvaluationShared
- move evaluation logic under Core/Language/Dictionary/Evaluation
- update Xcode project file references/groups/build phase entries for new layout
- remove legacy DictionaryMatcher+CandidateEvaluator and +SplitJoinEvaluator files
- update CODEMAP and ENGINEERING docs to reflect the new evaluation architecture
